# backend/openapi.yaml
openapi: 3.1.0
info:
  title: CC-Anywhere API
  description: Claude Code SDK HTTP Server API
  version: 0.4.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.cc-anywhere.com
    description: Production server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
          required:
            - code
            - message
      required:
        - error
    Task:
      type: object
      properties:
        taskId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        instruction:
          type: string
        result:
          type: object
          nullable: true
        error:
          type: object
          nullable: true
          properties:
            message:
              type: string
            code:
              type: string
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        workingDirectory:
          type: string
          nullable: true
      required:
        - taskId
        - status
        - instruction
        - createdAt
    CreateTaskRequest:
      type: object
      properties:
        instruction:
          type: string
          minLength: 1
        context:
          type: object
          properties:
            workingDirectory:
              type: string
            files:
              type: array
              items:
                type: string
        options:
          type: object
          properties:
            timeout:
              type: integer
              minimum: 1000
              maximum: 3600000
            async:
              type: boolean
            allowedTools:
              type: array
              items:
                type: string
            sdk:
              type: object
              description: Claude Code SDK specific options
              properties:
                maxTurns:
                  type: integer
                  minimum: 1
                  maximum: 50
                  description: Maximum conversation turns
                resumeSession:
                  type: string
                  description: Claude Code CLI session ID to resume
                  example: "69647d9d-d1a3-4924-8dc2-e9c558007a4b"
                continueFromTaskId:
                  type: string
                  description: Continue from a previous task using its SDK session ID
                permissionMode:
                  type: string
                  enum: [ask, allow, deny, acceptEdits, bypassPermissions, plan]
                  description: Permission mode for tool execution
                systemPrompt:
                  type: string
                  description: Custom system prompt
                allowedTools:
                  type: array
                  items:
                    type: string
                  description: List of allowed tools for SDK
                disallowedTools:
                  type: array
                  items:
                    type: string
                  description: List of disallowed tools for SDK
        continuationOptions:
          type: object
          properties:
            taskId:
              type: string
              format: uuid
      required:
        - instruction
    BatchTaskRequest:
      type: object
      properties:
        instruction:
          type: string
          minLength: 1
        repositories:
          type: array
          items:
            type: string
          minItems: 1
        options:
          type: object
          properties:
            timeout:
              type: integer
            concurrency:
              type: integer
              minimum: 1
              maximum: 10
      required:
        - instruction
        - repositories
...
    TaskGroupExecutionRequest:
      type: object
      properties:
        name:
          type: string
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskGroupTask'
          minItems: 1
          maxItems: 50
        execution:
          type: object
          properties:
            mode:
              type: string
              enum: [sequential, parallel, mixed]
            continueSession:
              type: boolean
              enum: [true]
              description: Always true - all tasks in a group share the same session
            continueOnError:
              type: boolean
              default: false
            timeout:
              type: integer
              minimum: 1000
              maximum: 600000
            permissionMode:
              type: string
              enum: [default, acceptEdits, bypassPermissions, plan, ask, allow, deny]
              default: bypassPermissions
              description: Permission mode for Claude Code SDK execution
          required:
            - mode
            - continueSession
        context:
          type: object
          properties:
            workingDirectory:
              type: string
            repositoryPath:
              type: string
          nullable: true
      required:
        - name
        - tasks
        - execution
    TaskGroupExecutionResponse:
      type: object
      properties:
        success:
          type: boolean
        groupId:
          type: string
        message:
          type: string
        totalTasks:
          type: integer
        executionPlan:
          type: array
          items:
            type: object
            properties:
              tasks:
                type: array
                items:
                  $ref: '#/components/schemas/TaskGroupTask'
              parallel:
                type: boolean
      required:
        - success
        - groupId
        - message
        - totalTasks
        - executionPlan
    TaskGroupStatus:
      type: object
      properties:
        groupId:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        sessionId:
          type: string
          nullable: true
        completedTasks:
          type: integer
        totalTasks:
          type: integer
        progress:
          type: integer
          minimum: 0
          maximum: 100
        tasks:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              status:
                type: string
                enum: [pending, running, completed, failed]
              error:
                type: string
                nullable: true
              startedAt:
                type: string
                format: date-time
                nullable: true
              completedAt:
                type: string
                format: date-time
                nullable: true
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - groupId
        - status
        - completedTasks
        - totalTasks
        - progress
        - tasks
        - startedAt
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - ApiKeyAuth: []
paths:
...
  /api/tasks:
    get:
      tags:
        - tasks
      summary: List all tasks
      operationId: listTasks
      parameters:
        - name: status
          in: query
          description: Filter by task status
          required: false
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
        - name: repository
          in: query
          description: Filter by repository
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of tasks to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of tasks to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Task list retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
                required:
                  - tasks
                  - total
                  - limit
                  - offset
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - tasks
      summary: Create a new task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/tasks/{taskId}:
    get:
      tags:
        - tasks
      summary: Get a specific task
      operationId: getTask
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - tasks
      summary: Cancel a task
      operationId: cancelTask
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  taskId:
                    type: string
                  status:
                    type: string
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/batch/tasks:
    post:
      tags:
        - batch
      summary: Create batch tasks
      operationId: createBatchTasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchTaskRequest'
      responses:
        '201':
          description: Batch tasks created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  groupId:
                    type: string
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  total:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/batch/tasks/{groupId}/status:
    get:
      tags:
        - batch
      summary: Get batch task status
      operationId: getBatchTaskStatus
      parameters:
        - name: groupId
          in: path
          required: true
          description: Batch group ID
          schema:
            type: string
      responses:
        '200':
          description: Batch status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  groupId:
                    type: string
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  summary:
                    type: object
                    properties:
                      total:
                        type: integer
                      pending:
                        type: integer
                      running:
                        type: integer
                      completed:
                        type: integer
                      failed:
                        type: integer
                      cancelled:
                        type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/queue/tasks:
    post:
      tags:
        - queue
      summary: Add task to queue
      operationId: addTaskToQueue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task added to queue successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  taskId:
                    type: string
                  position:
                    type: integer
                  status:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
        - queue
      summary: Get queued tasks
      operationId: getQueuedTasks
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by status
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
        - name: limit
          in: query
          required: false
          description: Maximum number of tasks to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          description: Number of tasks to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Queued tasks retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/queue/tasks/{taskId}:
    get:
      tags:
        - queue
      summary: Get queued task details
      operationId: getQueuedTask
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Queued task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - queue
      summary: Remove task from queue
      operationId: removeTaskFromQueue
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task removed from queue successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  taskId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/task-groups/execute:
    post:
      tags:
        - task-groups
      summary: Execute a task group with dependency management
      operationId: executeTaskGroup
      description: |
        Execute multiple tasks with dependency management and session continuity.
        Tasks can run sequentially, in parallel, or in a mixed mode based on dependencies.
        All tasks in a group share the same session for context continuity.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskGroupExecutionRequest'
      responses:
        '200':
          description: Task group execution started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskGroupExecutionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/task-groups/{groupId}/status:
    get:
      tags:
        - task-groups
      summary: Get task group execution status
      operationId: getTaskGroupStatus
      description: Get the current status and progress of a task group execution
      security:
        - ApiKeyAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          description: Task group ID
          schema:
            type: string
      responses:
        '200':
          description: Task group status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskGroupStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/task-groups/{groupId}:
    delete:
      tags:
        - task-groups
      summary: Cancel task group execution
      operationId: cancelTaskGroup
      description: Cancel a running task group execution
      security:
        - ApiKeyAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          description: Task group ID
          schema:
            type: string
      responses:
        '200':
          description: Task group cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/task-groups/cleanup:
    post:
      tags:
        - task-groups
      summary: Cleanup completed task groups
      operationId: cleanupTaskGroups
      description: Remove completed task groups from memory
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Task groups cleaned up successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'